Fixed a bug caused by non-unique wiki folder names
In the template code I used the name of the wiki or the folder to ID
the menu - but this caused problems where I have used the same folder
as a sub-folder of another folder...duplicate IDs in the html that
then prevents the menu click func from working.

Added an ID to the page that gets passed to the web and then use that
instead of the ID.  Better approach as I can now explicitely set the ID
rather than having random web code setting it.